Add files via upload
The menu options on the vending machine can now be shown or hidden with a function called ToggleMenu that toggles the onMenu state. This state changing dynamically changes the class of the div containing the menu options and the div containing the GoBack component so that if we are on a snack's url path, the only way to go back to the vending machine's menu is to click the Go Back link.